test: GameServiceTest 클래스, Provider 클래스 추출
